requirements.txt
diffusers>=0.30.3
accelerate>=0.34.2
transformers==4.46.2
numpy==1.26.0
torch==2.4.0
torchvision>=0.19.0
sentencepiece>=0.2.0
SwissArmyTransformer>=0.4.12
gradio>=4.44.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
openai>=1.45.05
moviepy>=1.0.3
pillow==9.5.0
scikit-video
peft
decord
blosc
bitsandbytes==0.45.2
timm
open-clip-torch
lightning
psutil
GPUtil
tabulate
pympler
objgraph
graphviz
torch-fidelity
ultralytics
opencv-python
git+https://github.com/openai/CLIP.git